refactor(provenance): isolate the abstract claim-control bases in a model_bases leaf
Move ClaimControlledModel and its Linkable* subclasses out of the provenance
domain into an apps.provenance.model_bases package, pinned as the bottom layer
of the provenance internal-stack import contract. The bases import only
apps.core, so the layer is a true leaf: a model can become claim-controlled by
inheriting it without depending on the provenance domain (Claim, ChangeSet,
Source) that builds on top.

Behavior-preserving and migration-free — the bases are abstract, so relocating
them keeps every (app_label, model_name) key, and external callers are
unchanged because apps.provenance.models still re-exports them.
