+++ backend/apps/provenance/model_bases/__init__.py
@@ -1,0 +1,18 @@
+"""The abstract claim-control base layer — see :mod:`.models`.
+
+Re-exports the bases so callers import ``from apps.provenance.model_bases import
+…`` (and the legacy ``from apps.provenance.models import …`` re-export keeps
+working) without reaching into the submodule.
+"""
+
+from .models import (
+    ClaimControlledModel,
+    LinkableClaimModel,
+    LinkableLifecycleClaimModel,
+)
+
+__all__ = [
+    "ClaimControlledModel",
+    "LinkableClaimModel",
+    "LinkableLifecycleClaimModel",
+]
